+++ integration/solana/balances.sol
@@ -1,0 +1,5 @@
+contract balances {
+	function get_balance(address addr) public view returns (uint64) {
+		return addr.balance;
+	}
+}