+++ tests/solana_tests/mappings.rs
@@ -33,6 +33,7 @@ fn simple_mapping() {
             ],
             &[],
             0,
+            None,
         );
     }
 
@@ -42,6 +43,7 @@ fn simple_mapping() {
             &[Token::Uint(ethereum_types::U256::from(102 + i))],
             &[],
             0,
+            None,
         );
 
         assert_eq!(
@@ -55,6 +57,7 @@ fn simple_mapping() {
         &[Token::Uint(ethereum_types::U256::from(101))],
         &[],
         0,
+        None,
     );
 
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(0))]);
@@ -64,6 +67,7 @@ fn simple_mapping() {
         &[Token::Uint(ethereum_types::U256::from(104))],
         &[],
         0,
+        None,
     );
 
     for i in 0..10 {
@@ -72,6 +76,7 @@ fn simple_mapping() {
             &[Token::Uint(ethereum_types::U256::from(102 + i))],
             &[],
             0,
+            None,
         );
 
         if 102 + i != 104 {
@@ -122,7 +127,7 @@ fn less_simple_mapping() {
         &[
             Token::Uint(ethereum_types::U256::from(12313132131321312311213131u128)),
             Token::String(String::from("This is a string which should be a little longer than 32 bytes so we the the abi encoder")),
-        ], &[], 0
+        ], &[], 0, None
     );
 
     vm.function(
@@ -133,6 +138,7 @@ fn less_simple_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
     let returns = vm.function(
@@ -142,6 +148,7 @@ fn less_simple_mapping() {
         ))],
         &[],
         0,
+        None,
     );
 
     assert_eq!(
@@ -190,7 +197,7 @@ fn string_mapping() {
         &[
             Token::String(String::from("a")),
             Token::String(String::from("This is a string which should be a little longer than 32 bytes so we the the abi encoder")),
-        ], &[],0
+        ], &[],0, None
     );
 
     vm.function(
@@ -201,9 +208,10 @@ fn string_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
-    let returns = vm.function("get", &[Token::String(String::from("a"))], &[], 0);
+    let returns = vm.function("get", &[Token::String(String::from("a"))], &[], 0, None);
 
     assert_eq!(
         returns,
@@ -238,6 +246,7 @@ fn mapping_in_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
     let returns = vm.function(
@@ -248,6 +257,7 @@ fn mapping_in_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
     assert_eq!(returns, vec![Token::FixedBytes(vec![0x98])]);
@@ -260,6 +270,7 @@ fn mapping_in_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
     assert_eq!(returns, vec![Token::FixedBytes(vec![0])]);
@@ -272,6 +283,7 @@ fn mapping_in_mapping() {
         ],
         &[],
         0,
+        None,
     );
 
     assert_eq!(returns, vec![Token::FixedBytes(vec![0])]);
@@ -314,7 +326,7 @@ fn sparse_array() {
         &[
             Token::Uint(ethereum_types::U256::from(909090909)),
             Token::String(String::from("This is a string which should be a little longer than 32 bytes so we the the abi encoder")),
-        ], &[],0
+        ], &[],0, None
     );
 
     vm.function(
@@ -325,6 +337,7 @@ fn sparse_array() {
         ],
         &[],
         0,
+        None,
     );
 
     let returns = vm.function(
@@ -332,6 +345,7 @@ fn sparse_array() {
         &[Token::Uint(ethereum_types::U256::from(909090909))],
         &[],
         0,
+        None,
     );
 
     assert_eq!(
@@ -380,7 +394,7 @@ fn massive_sparse_array() {
         &[
             Token::Uint(ethereum_types::U256::from(786868768768678687686877u128)),
             Token::String(String::from("This is a string which should be a little longer than 32 bytes so we the the abi encoder")),
-        ], &[],0
+        ], &[],0, None
     );
 
     vm.function(
@@ -391,6 +405,7 @@ fn massive_sparse_array() {
         ],
         &[],
         0,
+        None,
     );
 
     let returns = vm.function(
@@ -400,6 +415,7 @@ fn massive_sparse_array() {
         ))],
         &[],
         0,
+        None,
     );
 
     assert_eq!(