+++ tests/solana_tests/returns.rs
@@ -31,19 +31,19 @@ fn return_single() {
     );
     vm.constructor("foo", &[], 0);
 
-    let returns = vm.function("f", &[], &[], 0);
+    let returns = vm.function("f", &[], &[], 0, None);
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(2)),]);
 
-    let returns = vm.function("g", &[], &[], 0);
+    let returns = vm.function("g", &[], &[], 0, None);
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(3)),]);
 
-    let returns = vm.function("h", &[], &[], 0);
+    let returns = vm.function("h", &[], &[], 0, None);
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(2)),]);
 
-    let returns = vm.function("i", &[], &[], 0);
+    let returns = vm.function("i", &[], &[], 0, None);
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(24)),]);
 
-    let returns = vm.function("j", &[], &[], 0);
+    let returns = vm.function("j", &[], &[], 0, None);
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(5)),]);
 }
 
@@ -59,7 +59,7 @@ fn return_ternary() {
     );
 
     vm.constructor("foo", &[], 0);
-    let returns = vm.function("f", &[], &[], 0);
+    let returns = vm.function("f", &[], &[], 0, None);
 
     assert_eq!(
         returns,
@@ -79,7 +79,7 @@ fn return_ternary() {
     );
 
     vm.constructor("foo", &[], 0);
-    let returns = vm.function("f", &[], &[], 0);
+    let returns = vm.function("f", &[], &[], 0, None);
 
     assert_eq!(
         returns,
@@ -200,9 +200,9 @@ fn return_nothing() {
     );
 
     vm.constructor("foo", &[], 0);
-    let _returns = vm.function("strange", &[], &[], 0);
-    let _returns = vm.function("inc", &[], &[], 0);
-    let returns = vm.function("get", &[], &[], 0);
+    let _returns = vm.function("strange", &[], &[], 0, None);
+    let _returns = vm.function("inc", &[], &[], 0, None);
+    let returns = vm.function("get", &[], &[], 0, None);
 
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(2)),]);
 
@@ -230,8 +230,8 @@ fn return_nothing() {
     );
 
     vm.constructor("foo", &[], 0);
-    let _returns = vm.function("f", &[], &[], 0);
-    let returns = vm.function("get", &[], &[], 0);
+    let _returns = vm.function("f", &[], &[], 0, None);
+    let returns = vm.function("get", &[], &[], 0, None);
 
     assert_eq!(returns, vec![Token::Uint(ethereum_types::U256::from(5)),]);
 }
@@ -252,7 +252,7 @@ fn return_function() {
     );
 
     vm.constructor("foo", &[], 0);
-    let returns = vm.function("f", &[], &[], 0);
+    let returns = vm.function("f", &[], &[], 0, None);
 
     assert_eq!(
         returns,
@@ -276,7 +276,7 @@ fn return_function() {
     );
 
     vm.constructor("foo", &[], 0);
-    let returns = vm.function("f", &[], &[], 0);
+    let returns = vm.function("f", &[], &[], 0, None);
 
     assert_eq!(
         returns,